Refactor buildid based fix directory detection.
`ls -d */` although nice and simple fails when no directories are present
which is true in the case of just a global app fix. Since */ is expanded by
bash and is neccessary to skip . and such there is not really a way to fix
this using ls. Find, although a longer command, can achieve the desired
result.

Additionally, starting from the largest directory will typically be faster.